Split out common JSON-RPC server code.
This adds back a jsonrpc.py file, holding common code for JSON-RPC servers
as they are used both for the WatchTowerServer and the core Daemon.  By doing
so, we avoid duplicating that code (even if it is not so much), and ease
possible future maintenance.
